+++ rust_core/src/parsing.rs
@@ -18,6 +18,7 @@ pub enum SupportedLanguage {
     Cpp,
     CSharp,
     Sql,
+    Php,
 }
 
 impl SupportedLanguage {
@@ -34,6 +35,7 @@ impl SupportedLanguage {
             "cpp" | "cc" | "cxx" | "hpp" | "h" | "hxx" | "hh" => Some(SupportedLanguage::Cpp),
             "cs" => Some(SupportedLanguage::CSharp),
             "sql" => Some(SupportedLanguage::Sql),
+            "php" => Some(SupportedLanguage::Php),
             _ => None,
         }
     }
@@ -51,6 +53,7 @@ impl SupportedLanguage {
             SupportedLanguage::Cpp => tree_sitter_cpp::LANGUAGE.into(),
             SupportedLanguage::CSharp => tree_sitter_c_sharp::LANGUAGE.into(),
             SupportedLanguage::Sql => tree_sitter_sequel::LANGUAGE.into(),
+            SupportedLanguage::Php => tree_sitter_php::LANGUAGE_PHP.into(),
         }
     }
 
@@ -130,6 +133,14 @@ impl SupportedLanguage {
                 // SQL functions and procedures
                 r#"
                 (create_function) @function
+                "#
+            }
+            SupportedLanguage::Php => {
+                r#"
+                (function_definition
+                  name: (name) @name
+                  parameters: (formal_parameters) @params
+                  body: (compound_statement) @body) @function
                 "#
             }
         }
@@ -228,6 +239,20 @@ impl SupportedLanguage {
                 ]
                 "#
             }
+            SupportedLanguage::Php => {
+                // PHP classes, interfaces, and traits
+                r#"
+                [(class_declaration
+                  name: (name) @name
+                  body: (declaration_list) @body)
+                 (interface_declaration
+                  name: (name) @name
+                  body: (declaration_list) @body)
+                 (trait_declaration
+                  name: (name) @name
+                  body: (declaration_list) @body)] @class
+                "#
+            }
         }
     }
 }
@@ -314,6 +339,7 @@ impl CodeParser {
             SupportedLanguage::Cpp,
             SupportedLanguage::CSharp,
             SupportedLanguage::Sql,
+            SupportedLanguage::Php,
         ] {
             let mut parser = Parser::new();
             parser